minesweeper_game/requirements.txt
tkinter
playsound
pillow
